+++ History.md
@@ -1,5 +1,7 @@
 ## 2.5.2 *
 - Moved to LLVM 8.0
+- Added Asparagus and Pear models to the gallery, removed Elderflower
+- Added support for image preprocessing metadata and image normalization defaults for pytorch + imagenet
 
 ## 2.5.1
 - Fixed API wrapper for PortMemoryLayout to produce correct extent